Added the common  authentication routes
