+++ frontend/src/App.jsx
@@ -12,6 +12,7 @@ import Signup from "./components/Auth/signup/signup";
 import MyDaily_status_edit from "./components/Attendance/pages/MyDailyStatus/MyDaily_status_edit";
 import { ThemeProvider, createTheme } from "@mui/material/styles";
 import CssBaseline from "@mui/material/CssBaseline";
+import { AuthenticationRoutes } from "./components/Auth/authenticatin";
 const darkTheme = createTheme({
   palette: {
     mode: "dark",
@@ -30,23 +31,28 @@ function App() {
         <Router>
           <Routes>
             <Route path="/login" element={<Login />} />
-            <Route path="/" element={<Dashboard />} />
-            <Route path="/daily_status_updates" element={<MyDailyStatus />} />
-            <Route
-              path="/daily_status_updates_details/"
-              element={<MyDailyStatusNew_id />}
-            />
-
-            <Route
-              path="/daily_status_updates_details_edit"
-              element={<MyDaily_status_edit />}
-            />
-            <Route path="/send_daily_status" element={<SendMyDailyStatus />} />
-            <Route path="/my_leave" element={<MyLeave />} />
-            <Route path="/holidays" element={<Holidays />} />
             <Route path="/forgot-password" element={<ForgotPassword />} />
             {role === "admin" && <Route path="/Signup" element={<Signup />} />}
-            <Route path="*" element={<NotFound />} />
+            
+              <Route path="/" element={<AuthenticationRoutes><Dashboard /> </AuthenticationRoutes>} />
+              <Route path="/daily_status_updates" element={<MyDailyStatus />} />
+              <Route
+                path="/daily_status_updates_details/"
+                element={<MyDailyStatusNew_id />}
+              />
+
+              <Route
+                path="/daily_status_updates_details_edit"
+                element={<MyDaily_status_edit />}
+              />
+              <Route
+                path="/send_daily_status"
+                element={<SendMyDailyStatus />}
+              />
+              <Route path="/my_leave" element={<MyLeave />} />
+              <Route path="/holidays" element={<Holidays />} />
+              <Route path="*" element={<NotFound />} />
+           
           </Routes>
         </Router>
       </ThemeProvider>
