+++ frontend/src/components/Attendance/pages/Dashboard/Dashboard.jsx
@@ -1,20 +1,8 @@
-import React, { useEffect } from "react";
-import { useNavigate } from "react-router-dom";
+import React from "react";
 import Header from "../../Header/Header";
 import "../../dashboard.css";
 
 const Dashboard = () => {
-  const navigate = useNavigate();
-  useEffect(() => {
-    const localToken = localStorage.getItem("jwtToken");
-    if (!localToken) {
-      navigate("/login");
-      return;
-    } else {
-      navigate("/");
-      return;
-    }
-  }, []);
   return (
     <>
       <Header />
