+++ frontend/src/components/Attendance/pages/MyDailyStatus/MyDailyStatusNew_id.jsx
@@ -1,19 +1,10 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import Layout from "../../../Layout/Layout";
-import Header from "../../Header/Header";
 import "../../dashboard.css";
 export const MyDailyStatusNew_id = () => {
   const { state } = useLocation();
   const data = state.item;
-  const navigate = useNavigate();
-  useEffect(() => {
-    const token = localStorage.getItem("jwtToken");
-    if (!token) {
-      navigate("/login");
-      return;
-    }
-  }, []);
   return (
     <>
       <Layout>
